Revert "Avoid wm<->am deadlock when checking uri permission"
This reverts commit f0ee42681a2298fa4d6567cfbbb6cab6e214ec04.

Bug: 157106728
Bug: 115619667
Reason for revert:

Looked into this some with the Camera test case in b/157106728. The wait()/notify strategy isn't going to work here because we have a 3-lock system vs. 2. We have the WM, UGM, and AM locks involved. Both the WM and UGM lock are held in the call to ActivityManagerInternal#checkContentProviderUriPermission and in the subsequence call that ActivityManagerService#checkContentProviderUriPermission AMS calls back into WMS and UGMS for various reasons which requires their locks...wait()/notify strategy allows us to break out of 1 lock, but things get really messed up if we try to use that strategy to break out of 2 locks.

Change-Id: Ic0f30f6c1b24334105a00790488d2758390e85a1