Extract verification logic out of InstallParams
InstallParams is now solely used for installing packages. If we call
installStage() method then package will be installed without
verification. This function is now used during movePackage to bypass
verification. In future, we can use this for staged installation to
bypass verification during boot.

The extracted logic is placed into VerificationParams. Once verification
is completed, we no longer continue to install phase. We return the
result to caller and allow the caller to decide how the flow will
progress.

Bug: 159331446
Test: atest StagedInstallTest
Test: atest GtsSecurityHostTestCases
Test: atest CtsAtomicInstallTestCases
Change-Id: I4b831a3eb194d4cf02e0d65adc9f0dcdb1628a0e