Show secondary home handle in immersive apps
If user is quickswitching, we show the secondary
home handle to let them know they can continue
quickswitching.
Also increased the secondary handle height region
to match that of the regular nav bar fragment,
the old height was causing too many false positives
for recents to think a touch was inside the app.

Bug: 158677967
Change-Id: I3a0e5bfa228eeb376a59ae4e82ed47589cbc51b0